MV-Adapter/requirements.txt
controlnet_aux
diffusers
transformers
peft
numpy
huggingface_hub
accelerate
opencv-python
safetensors
pillow
omegaconf
trimesh
einops
gradio
timm
kornia
scikit-image
sentencepiece